Basic Function
Added 6 enemies, 1 player.
Made enemies move across the screen over and over again.
Set key control of the player, and prevent it going off the canvas.
Set the player to the starting point when it reaches the river.
Set the player to the starting point when it collide with the enemies.